Move Apple Speech tests into focused targets
