+++ TypeWhisperPluginSDK/Plugins/SpeechAnalyzerPlugin/Tests/AppleSpeechModelSelectionTests.swift
